requirements.txt
Django==1.8.7
django-cors-headers==1.1.10
git+git://github.com/ellmetha/django-machina.git
djangorestframework==3.3.1
djangorestframework-jwt==1.7.2
django-webpack-loader==0.2.2
Unipath==1.1
Whoosh==2.7.0
